refactor(native): extract window geometry
